requirements.txt
django
drf-yasg
djangorestframework
celery
redis
gunicorn
pillow
psycopg2-binary
python-decouple
django-filter
djangorestframework-simplejwt
django-ckeditor
